Refactor story brief schema validation into focused modules
